[ github actions ] adding conditional steps based on running OS
This was needed because the CI tried to use macOS signing certificates to sign the Windows build and failed ofc. So I have different steps now depending on the operating system. Why electron-builder decided to use the same environment variable names for both windows and mac when their certificates are not compatible is beyond me.